[#66] refactor : 폴더 정리 및 머지 오류 수정
